requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
pyspark==3.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data extraction
yfinance==0.2.28
newsapi-python==0.2.7
tweepy==4.14.0
requests==2.31.0
beautifulsoup4==4.12.2

# Sentiment analysis
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.36.2
torch==2.1.2

# Technical analysis
ta-lib==0.4.28
talib-binary==0.4.19

# Web framework and visualization
streamlit==1.29.0
flask==3.0.0
plotly==5.17.0
dash==2.16.1
bokeh==3.3.2

# Database
python-dotenv==1.0.0
alembic==1.13.1

# Utilities
schedule==1.2.1
click==8.1.7
tqdm==4.66.1
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
jupyter==1.0.0
black==23.12.1
flake8==6.1.0
